splice: make windows and darwin happy
