+++ splice/pair_windows.go
